requirements.txt
azure-graphrbac
pyyaml
oauth2client
click
